Small edits to the introduction
I know where you're going with this, but it is currently pretty flammable stuff. I think that citing cases of all of these points will help mitigate the story that you're weaving. I would also run this by Outlier at some point, maybe - it's kind of hard to come down on VCs when we have VC money. Something to think about.

Happy to work through this all with you in person. I think a nice session outlining how this chapter should go would be a good idea for the Azores.

I am going to move on to the next chapter. This one clearly isn't done yet. That's OK!